Update UMS.conf for compatibility with ext/int subtitles
Tested with a Sony Bravia KDL-46EX720. Which can be particular about how
the subtitle streams are handled.

With this setup, the default mencoder stream successfully supports
embedded subtitles and external SRT subs on this renderer.

Mencoder is also set to force a 1920/1080 resolution to compensate for
this particular TV's rendering of small resolutions. If all your content
is 16/9, this isn't a problem.